requirements.txt
python-jobspy
python-dotenv
py-cord
validators